# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tandava Tour Company API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface Package {
  id: number;
  title: string;
  destination: string;
  duration: string;
  /** Starting price in INR */
  startingPrice: number;
  description: string;
  includedServices: string[];
  imageUrl: string;
  highlights?: string[];
  active: boolean;
  sortOrder: number;
  createdAt: string;
}

export interface PackageInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  destination: string;
  /** @minLength 1 */
  duration: string;
  /** @minimum 0 */
  startingPrice: number;
  description: string;
  includedServices: string[];
  imageUrl: string;
  highlights?: string[];
  active?: boolean;
  sortOrder?: number;
}

export interface Vehicle {
  id: number;
  name: string;
  vehicleType: string;
  seatingCapacity: number;
  airConditioned: boolean;
  musicSystem: boolean;
  imageUrl: string;
  description: string;
  features?: string[];
  active: boolean;
  sortOrder: number;
  createdAt: string;
}

export interface VehicleInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  vehicleType: string;
  /** @minimum 1 */
  seatingCapacity: number;
  airConditioned: boolean;
  musicSystem: boolean;
  imageUrl: string;
  description: string;
  features?: string[];
  active?: boolean;
  sortOrder?: number;
}

export interface GalleryImage {
  id: number;
  title?: string;
  caption?: string;
  imageUrl: string;
  sortOrder: number;
  createdAt: string;
}

export interface GalleryImageInput {
  title?: string;
  caption?: string;
  imageUrl: string;
  sortOrder?: number;
}

export interface Review {
  id: number;
  authorName: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  comment: string;
  location?: string;
  avatarUrl?: string;
  createdAt: string;
}

export interface ReviewInput {
  /** @minLength 1 */
  authorName: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  /** @minLength 1 */
  comment: string;
  location?: string;
  avatarUrl?: string;
}

export interface Settings {
  id: number;
  companyName: string;
  tagline: string;
  phonePrimary: string;
  phoneSecondary: string;
  whatsappNumber: string;
  email: string;
  location: string;
  instagramUrl: string;
  facebookUrl: string;
  youtubeUrl: string;
  aboutText: string;
  updatedAt: string;
}

export interface SettingsInput {
  companyName?: string;
  tagline?: string;
  phonePrimary?: string;
  phoneSecondary?: string;
  whatsappNumber?: string;
  email?: string;
  location?: string;
  instagramUrl?: string;
  facebookUrl?: string;
  youtubeUrl?: string;
  aboutText?: string;
}

export interface InstagramPost {
  id: number;
  title: string;
  caption?: string;
  imageUrl: string;
  postUrl: string;
  displayOrder: number;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface InstagramPostInput {
  /** @minLength 1 */
  title: string;
  caption?: string;
  /** @minLength 1 */
  imageUrl: string;
  /** @minLength 1 */
  postUrl: string;
  displayOrder?: number;
  isActive?: boolean;
}

export interface AdminLoginRequest {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}

export interface AdminSession {
  username: string;
  authenticated: boolean;
}

export interface DashboardStats {
  totalPackages: number;
  activePackages: number;
  totalVehicles: number;
  activeVehicles: number;
  totalGalleryImages: number;
  totalReviews: number;
  averageRating: number;
}

export type ListPackagesParams = {
  includeInactive?: boolean;
};

export type ListVehiclesParams = {
  includeInactive?: boolean;
};

export type ListInstagramPostsParams = {
  includeInactive?: boolean;
};