# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tandava Tour Company API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /packages:
    get:
      tags: [packages]
      operationId: listPackages
      summary: List active tour packages
      parameters:
        - name: includeInactive
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Package"
  /admin/packages:
    post:
      tags: [packages, admin]
      operationId: createPackage
      summary: Create a tour package (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/packages/{id}:
    put:
      tags: [packages, admin]
      operationId: updatePackage
      summary: Update a tour package (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      tags: [packages, admin]
      operationId: deletePackage
      summary: Delete a tour package (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /admin/gallery:
    post:
      tags: [gallery, admin]
      operationId: createGalleryImage
      summary: Add gallery image (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryImageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryImage"
  /admin/gallery/{id}:
    delete:
      tags: [gallery, admin]
      operationId: deleteGalleryImage
      summary: Delete gallery image (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /instagram-posts:
    get:
      tags: [instagram]
      operationId: listInstagramPosts
      summary: List Instagram posts displayed on the public site
      parameters:
        - name: includeInactive
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Instagram posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InstagramPost"
  /admin/instagram-posts:
    post:
      tags: [instagram, admin]
      operationId: createInstagramPost
      summary: Create an Instagram post (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstagramPostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstagramPost"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/instagram-posts/{id}:
    put:
      tags: [instagram, admin]
      operationId: updateInstagramPost
      summary: Update an Instagram post (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstagramPostInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstagramPost"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      tags: [instagram, admin]
      operationId: deleteInstagramPost
      summary: Delete an Instagram post (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admin/login:
    post:
      tags: [admin]
      operationId: adminLogin
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /admin/me:
    get:
      tags: [admin]
      operationId: adminMe
      summary: Get current admin session
      responses:
        "200":
          description: Session info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/dashboard-stats:
    get:
      tags: [admin]
      operationId: adminDashboardStats
      summary: Aggregate stats for the admin dashboard
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Package:
      type: object
      required:
        - id
        - title
        - destination
        - duration
        - startingPrice
        - description
        - includedServices
        - imageUrl
        - active
        - sortOrder
        - createdAt
      properties:
        id:
          type: integer
        title:
          type: string
        destination:
          type: string
        duration:
          type: string
        startingPrice:
          type: integer
          description: Starting price in INR
        description:
          type: string
        includedServices:
          type: array
          items:
            type: string
        imageUrl:
          type: string
        highlights:
          type: array
          items:
            type: string
        active:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
    PackageInput:
      type: object
      required:
        - title
        - destination
        - duration
        - startingPrice
        - description
        - includedServices
        - imageUrl
      properties:
        title:
          type: string
          minLength: 1
        destination:
          type: string
          minLength: 1
        duration:
          type: string
          minLength: 1
        startingPrice:
          type: integer
          minimum: 0
        description:
          type: string
        includedServices:
          type: array
          items:
            type: string
        imageUrl:
          type: string
        highlights:
          type: array
          items:
            type: string
        active:
          type: boolean
        sortOrder:
          type: integer
...
    GalleryImage:
      type: object
      required:
        - id
        - imageUrl
        - sortOrder
        - createdAt
      properties:
        id:
          type: integer
        title:
          type: string
        caption:
          type: string
        imageUrl:
          type: string
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
    GalleryImageInput:
      type: object
      required:
        - imageUrl
      properties:
        title:
          type: string
        caption:
          type: string
        imageUrl:
          type: string
        sortOrder:
          type: integer
...
    InstagramPost:
      type: object
      required:
        - id
        - title
        - imageUrl
        - postUrl
        - displayOrder
        - isActive
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        title:
          type: string
        caption:
          type: string
        imageUrl:
          type: string
        postUrl:
          type: string
        displayOrder:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    InstagramPostInput:
      type: object
      required:
        - title
        - imageUrl
        - postUrl
      properties:
        title:
          type: string
          minLength: 1
        caption:
          type: string
        imageUrl:
          type: string
          minLength: 1
        postUrl:
          type: string
          minLength: 1
        displayOrder:
          type: integer
        isActive:
          type: boolean
    AdminLoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
    AdminSession:
      type: object
      required: [username, authenticated]
      properties:
        username:
          type: string
        authenticated:
          type: boolean
    DashboardStats:
      type: object
      required:
        - totalPackages
        - activePackages
        - totalVehicles
        - activeVehicles
        - totalGalleryImages
        - totalReviews
        - averageRating
      properties:
        totalPackages:
          type: integer
        activePackages:
          type: integer
        totalVehicles:
          type: integer
        activeVehicles:
          type: integer
        totalGalleryImages:
          type: integer
        totalReviews:
          type: integer
        averageRating:
          type: number