# lib/api-zod/src/generated/types/instagramPost.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tandava Tour Company API
 * OpenAPI spec version: 0.1.0
 */

export interface InstagramPost {
  id: number;
  title: string;
  caption?: string;
  imageUrl: string;
  postUrl: string;
  displayOrder: number;
  isActive: boolean;
  createdAt: Date;
  updatedAt: Date;
}