# lib/api-zod/src/generated/types/packageInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tandava Tour Company API
 * OpenAPI spec version: 0.1.0
 */

export interface PackageInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  destination: string;
  /** @minLength 1 */
  duration: string;
  /** @minimum 0 */
  startingPrice: number;
  description: string;
  includedServices: string[];
  imageUrl: string;
  highlights?: string[];
  active?: boolean;
  sortOrder?: number;
}